Add rackspace scheduler
Serializes builds

Weights by IOps, os_type, etc

Fuzzes the scheduling so we're not picking the exact same host
on every build...

Add the concept of spares.  You can configure a percentage of hosts to
reserve as spares ('spare_host_percentage' setting, the default is 10).

If there wouldn't be room to build an instance otherwise, we'll dip into
the spares.

This patch contains a number of other things that should hit trunk to
support some of the tests I've added.. as well as printing the weight
with a WeightedHost.  Might be a couple other things I'm forgetting.