Xen: Ensure snapshot is torn down on error.
This patch solves the problem of snapshots being left around if an
exception is generated in _wait_for_coalesce.

The solution is to separate snapshot from coalescing and vdi-chain
walking and then ensuring any code exceptions generated after
snapshotting trigger a cleanup.
If an exception is raised in _wait_for_coalesce,

Change-Id: I4c03243bc41f76d80d32934dd9506d1861951184