Ensure builds work with cells.
A recent change to upstream HEAD moved the location of newly-built
instance UUIDs from request_spec["instance_properties"]["uuid"] to
request_spec["instance_uuids"] and is now a list of UUIDs to be built.

While this feature should be supported in the future, the quick fix
is to just to continue to support building one instance at a time.

A future fix should be scheduled to ensure compatibility with
creating multiple instances in the same create request.

Change-Id: Ib3fa024585117439c084720ccf26ec067fe3dd8c